refactor(desktop): split BusinessHarnessView across Plan/Inspect/Operate (Group B)
Splits the 1100-line BusinessHarnessView monolith into three verb-specific
bodies for the entity/verb workspace model:

  * BusinessPlanBody    — identity header, seven-folder grid + per-folder
                          artifact panel, products/runtimes/external-systems/
                          access-tiers/AI-loop budget pointer rows.
  * BusinessInspectBody — index-session drift summary pills, AI-loop event
                          stream, read-only customers drift detection.
  * BusinessOperateBody — focused open-drifts list with apply buttons +
                          disabled "Run indexer loop now" placeholder
                          (wired in Group E).

Shared presentational helpers (DriftStatPill, DriftRow, FolderRow,
LoopEventLog, PointerRow, etc.) extracted to BusinessHarnessShared.tsx.
BusinessWorkspace now wires the three bodies to plan/inspect/operate
verbs; the legacy harness/BusinessHarnessView.tsx file is deleted.

No behavior change — same Catppuccin styling, same IPC routes
(detectCustomerDrift, applyCustomerDrift, listHarnessFolder,
recentHarnessLoopEvents, onHarnessLoopEvent). Drift detection still runs
on cold open of Inspect/Operate. Read-only Inspect surface points users
at Operate to apply drifts.

Verified: npx turbo run build --filter=@brainst0rm/desktop builds clean.
